+++ components/wysiwyg/widget-block/index.tsx
@@ -3,6 +3,7 @@ import { useFetchWidget } from 'hooks/widget';
 import WidgetBlock from './component';
 
 import type { WYSIWYGItem } from 'types/wysiwyg';
+import type { APIWidgetSpec } from 'types/widget';
 
 export interface WidgetBlockContainerProps {
   item: WYSIWYGItem;
@@ -18,7 +19,7 @@ const WidgetBlockContainer = ({ item }: WidgetBlockContainerProps): JSX.Element 
     },
   );
 
-  return <WidgetBlock widget={widget} />;
+  return <WidgetBlock widget={widget as APIWidgetSpec} />;
 };
 
 export default WidgetBlockContainer;