+++ layout/explore/explore-favorites/component.jsx
@@ -1,6 +1,4 @@
-import {
-  useMemo,
-} from 'react';
+import { useMemo } from 'react';
 import PropTypes from 'prop-types';
 
 // components
@@ -11,64 +9,49 @@ import ExploreDatasetsActions from 'layout/explore/explore-datasets/explore-data
 import CardPlaceholder from 'components/card-placeholder';
 
 // hooks
-import useFetchUserFavorites from 'hooks/favorite/fetch-favorites';
+import { useFetchUserFavorites } from 'hooks/favorite';
 import useFetchDatasets from 'hooks/dataset/fetch-datasets';
 
-const ExploreFavorites = ({
-  userToken,
-}) => {
-  const {
-    data: favorites,
-  } = useFetchUserFavorites([userToken], {
+const ExploreFavorites = ({ userToken }) => {
+  const { data: favorites } = useFetchUserFavorites([userToken], {
     initialData: [],
     initialStale: true,
   });
-  const favoriteDatasets = useMemo(() => favorites.filter(({ resourceType }) => resourceType === 'dataset'), [favorites]);
+  const favoriteDatasets = useMemo(
+    () => favorites.filter(({ resourceType }) => resourceType === 'dataset'),
+    [favorites],
+  );
   const datasetIds = useMemo(
     () => favoriteDatasets.map(({ resourceId }) => resourceId),
     [favoriteDatasets],
   );
-  const {
-    data: datasets,
-    isFetchedAfterMount,
-  } = useFetchDatasets({
-    ids: datasetIds.join(','),
-    includes: 'widget,metadata,layer,vocabulary',
-    env: process.env.NEXT_PUBLIC_ENVS_SHOW,
-  },
-  {
-    enabled: !!(datasetIds.length),
-    initialData: [],
-    initialStale: true,
-    keepPreviousData: true,
-  });
+  const { data: datasets, isFetchedAfterMount } = useFetchDatasets(
+    {
+      ids: datasetIds.join(','),
+      includes: 'widget,metadata,layer,vocabulary',
+      env: process.env.NEXT_PUBLIC_ENVS_SHOW,
+    },
+    {
+      enabled: !!datasetIds.length,
+      initialData: [],
+      initialStale: true,
+      keepPreviousData: true,
+    },
+  );
 
   return (
     <div className="c-explore-favorites">
-      {(!isFetchedAfterMount && !datasets.length) && (
-        <Spinner
-          isLoading
-          className="-light"
-        />
-      )}
-      <DatasetList
-        list={datasets}
-        actions={<ExploreDatasetsActions />}
-      />
-      {(isFetchedAfterMount && !datasets.length) && (
+      {!isFetchedAfterMount && !datasets.length && <Spinner isLoading className="-light" />}
+      <DatasetList list={datasets} actions={<ExploreDatasetsActions />} />
+      {isFetchedAfterMount && !datasets.length && (
         <div className="no-datasets">
           <CardPlaceholder />
           <div className="message">
             <h5>You currently have no favorite datasets</h5>
             <p>
-              To favorite a dataset or start a collection, click the
-              &nbsp;
-              <Icon
-                name="icon-star-full"
-                className="-star -small"
-              />
-              &nbsp;
-              on any dataset card.
+              To favorite a dataset or start a collection, click the &nbsp;
+              <Icon name="icon-star-full" className="-star -small" />
+              &nbsp; on any dataset card.
             </p>
           </div>
           <CardPlaceholder />