+++ layout/reset-password/index.js
@@ -1,9 +1,0 @@
-// Redux
-import { connect } from 'react-redux';
-
-import ResetPassword from './component';
-
-export default connect(
-  null,
-  null,
-)(ResetPassword);